New auth flow. New Generic credentials object.
Added support for a auth flow: Client Credentials.

To support that, added a new class "GenericCredentials" which
generalizes the different ways the new auth flow and the PKCE auth flow
return tokens.

Now the auth manager can operate on the auth flows, and the credentials
the auth flows return in a generic manner (through the interface, like
it should).