More refactoring.
Seperate concerns... loadTokens() was doing 2 things, validating the
cache, as much as it could without pinging spotify, and then
constructing a GenericCredentials object IF the cache was good.

That's 2 things. I made them each their own function.

I also eliminated the CacheState enum in favor of a boolean.
Moved some of the logging into TokenCache, which got rid of the need
for a lot of the enum states.